Refactored JFrameGraphTest to JFrameGraph
